% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kakarayan.R
\name{kakarayan_concordance}
\alias{kakarayan_concordance}
\title{Query concordance lines}
\usage{
kakarayan_concordance(client, q, language_id, ...)
}
\arguments{
\item{client}{Kakarayan client.}

\item{q}{Query text.}

\item{language_id}{Stable language identifier.}

\item{...}{Additional documented live API parameters.}
}
\description{
Query concordance lines
}
