% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kakarayan.R
\name{kakarayan_corpora}
\alias{kakarayan_corpora}
\title{Read the corpus catalogue}
\usage{
kakarayan_corpora(client)
}
\arguments{
\item{client}{Kakarayan client.}
}
\description{
Read the corpus catalogue
}
