% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kakarayan.R
\name{kakarayan_dictionary}
\alias{kakarayan_dictionary}
\title{Query dictionary candidates}
\usage{
kakarayan_dictionary(client, q, language_id, ...)
}
\arguments{
\item{client}{Kakarayan client.}

\item{q}{Query text.}

\item{language_id}{Stable language identifier.}

\item{...}{Additional documented live API parameters.}
}
\description{
Query dictionary candidates
}
