% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kakarayan.R
\name{kakarayan_frequencies}
\alias{kakarayan_frequencies}
\title{Query token frequencies}
\usage{
kakarayan_frequencies(client, language_id, ...)
}
\arguments{
\item{client}{Kakarayan client.}

\item{language_id}{Stable language identifier.}

\item{...}{Additional documented live API parameters.}
}
\description{
Query token frequencies
}
