% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kakarayan.R
\name{kakarayan_languages}
\alias{kakarayan_languages}
\title{Read the language catalogue}
\usage{
kakarayan_languages(client)
}
\arguments{
\item{client}{Kakarayan client.}
}
\description{
Read the language catalogue
}
