% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kakarayan.R
\name{kakarayan_search_manifest}
\alias{kakarayan_search_manifest}
\title{Read the static search manifest}
\usage{
kakarayan_search_manifest(client)
}
\arguments{
\item{client}{Kakarayan client.}
}
\description{
Read the static search manifest
}
