% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kakarayan.R
\name{kakarayan_search_shard}
\alias{kakarayan_search_shard}
\title{Read one allowlisted static search shard}
\usage{
kakarayan_search_shard(client, path)
}
\arguments{
\item{client}{Kakarayan client.}

\item{path}{Path from the search manifest.}
}
\description{
Read one allowlisted static search shard
}
